### 2017-03-18 #211 @whedon
[PRE REVIEW]: SaffronTree: Fast, reference-free pseudo-phylogenomic trees from reads or contigs
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/sanger-pathogens/saffrontree" target ="_blank">https://github.com/sanger-pathogens/saffrontree</a>
**Version:** v0.1.2
