### 2017-04-17 #243 @whedon
[REVIEW]: SaffronTree: Fast, reference-free pseudo-phylogenomic trees from reads or contigs
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/sanger-pathogens/saffrontree" target ="_blank">https://github.com/sanger-pathogens/saffrontree</a>
**Version:** v0.1.2
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.4964378.v1" target="_blank">10.6084/m9.figshare.4964378.v1</a>
