### 2017-04-24 #243 @brainstorm
Hi @andrewjpage! I've successfully installed the python packages locally in an Anaconda python3 environment, but there seems to be assumptions about where "KMC" is setup and the tests end up failing:

```
(joss) romanvgs-Air:saffrontree romanvg$ ./run_tests.sh
Tests should not be run pathpipe-farm3
..E..Running: kmc -k10 -fq -ci10 -cx200 -t1 /Users/romanvg/tmp/saffrontree/saffrontree/tests/data/kmcfastq/S_typhi_CT18_chromosome_1.fastq.gz /Users/romanvg/tmp/saffrontree/tmp_samplekmers_m5_6c8b0/fastq_kmers /Users/romanvg/tmp/saffrontree/tmp_samplekmers_m5_6c8b0
.F......FE
ERROR: test_kmc_command (KmcDetectVersion_test.TestKmcVersionDetect)
Traceback (most recent call last):
  File "/Users/romanvg/tmp/saffrontree/saffrontree/tests/KmcDetectVersion_test.py", line 11, in test_kmc_command
    k = KmcVersionDetect(False)
  File "/Users/romanvg/tmp/saffrontree/saffrontree/KmcVersionDetect.py", line 11, in __init__
    self.kmc_version = self.find_version()
  File "/Users/romanvg/tmp/saffrontree/saffrontree/KmcVersionDetect.py", line 21, in find_version
    kmc_output = check_output(["kmc"], universal_newlines=True)
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 336, in check_output
    **kwargs).stdout
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 403, in run
    with Popen(*popenargs, **kwargs) as process:
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 707, in __init__
    restore_signals, start_new_session)
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 1326, in _execute_child
    raise child_exception_type(errno_num, err_msg)
FileNotFoundError: [Errno 2] No such file or directory: 'kmc'

ERROR: test_fastq_files (SaffronTree_test.TestSaffronTree)
Traceback (most recent call last):
  File "/Users/romanvg/tmp/saffrontree/saffrontree/tests/SaffronTree_test.py", line 40, in test_fastq_files
    st = SaffronTree(options)
  File "/Users/romanvg/tmp/saffrontree/saffrontree/SaffronTree.py", line 37, in __init__
    self.kmc_major_version = KmcVersionDetect(self.verbose).major_version()
  File "/Users/romanvg/tmp/saffrontree/saffrontree/KmcVersionDetect.py", line 11, in __init__
    self.kmc_version = self.find_version()
  File "/Users/romanvg/tmp/saffrontree/saffrontree/KmcVersionDetect.py", line 21, in find_version
    kmc_output = check_output(["kmc"], universal_newlines=True)
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 336, in check_output
    **kwargs).stdout
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 403, in run
    with Popen(*popenargs, **kwargs) as process:
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 707, in __init__
    restore_signals, start_new_session)
  File "/Users/romanvg/.miniconda/envs/joss/lib/python3.6/subprocess.py", line 1326, in _execute_child
    raise child_exception_type(errno_num, err_msg)
FileNotFoundError: [Errno 2] No such file or directory: 'kmc'

FAIL: test_running_kmc_command (KmcFastq_test.TestKmcFastq)
Traceback (most recent call last):
  File "/Users/romanvg/tmp/saffrontree/saffrontree/tests/KmcFastq_test.py", line 46, in test_running_kmc_command
    self.assertTrue(os.path.exists(os.path.join(k.temp_working_dir,'fastq_kmers.kmc_pre')))
AssertionError: False is not true

FAIL: test_running_intersection_and_counting_kmers (KmcIntersect_test.TestKmcIntersect)
Traceback (most recent call last):
  File "/Users/romanvg/tmp/saffrontree/saffrontree/tests/KmcIntersect_test.py", line 64, in test_running_intersection_and_counting_kmers
    self.assertTrue(os.path.exists('results.kmc_pre'))
AssertionError: False is not true

Ran 15 tests in 0.051s

FAILED (failures=2, errors=2)
```

That is because after `source install_dependencies.sh`, KMC ends up under `build` directory:

```
$ find . -iname kmc
./build/kmc
```

Just a minor detail, the docker counterpart works fine!
