### 2017-04-27 #243 @andrewjpage
Hi @brainstorm,
Thanks for trying the software out. The install dependancies script is for travis, so I've renamed it to avoid confusion and I've added [installation instructions for Anaconda](https://github.com/sanger-pathogens/saffrontree/blob/master/README.md).
Andrew
