### 2017-05-01 #243 @brainstorm
Fantastic! Thanks for that @andrewjpage for that, it worked like a charm. I'll move on to the functionality aspects of the package, thanks!
