### 2017-05-01 #243 @brainstorm
Just a minor observation as I proceed with the review, the following online newick tree viewer states that trees should start with parenthesis:

http://trex.uqam.ca/index.php?action=newick

The generated trees from the example data begin with `[&R]` instead... also the online tool claims that the object length names should not exceed 50 characters.

FigTree does work fine though, but other alternative and tested viewers/libraries for newick on the FAQ would be great to have.
