### 2017-05-02 e994afc
note about newick viewing
