### 2017-05-03 #243 @arfon
@brainstorm - many thanks for your review here and to @biorelated for editing this one ✨

@andrewjpage - your submission is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00243 ⚡️ 🚀 💥

In other news, this is the 100th accepted JOSS paper 🎉! Nice to see this coming from the Sanger (I worked in the Production Software Group in ~2008/2009)
